failing codegen test case for tables > 440 columns
currently fails with

[info] Compiling 4 Scala sources to slick/slick-testkit/target/scala-2.11/test-classes...
[error] slick/slick-testkit/target/scala-2.11/src_managed/test/slick-codegen/slick/test/codegen/roundtrip/Tables.scala:550: object <none> is not a member of package scala
[error]       val data = ((x.column1, x.column2, x.column3, ...

which is a weird error message but means that we tried to pass more
22 elements into a tuple. That is the case for the outer tuple,
to which the generated code passes all groups, which are more than
22 here. We need to determine an appropriate nesting depth for the
given number of columns and then nest tuples that deep, rather than
just using depth 2.

(ideally we would code-generate HugeTable.scala but I failed to get
freemarker to work here. Low priority)